+++ src/lib/interfaces/index.ts
@@ -1,7 +1,7 @@
 export * from './column-filter-event';
 export * from './filter-change-event';
 export * from './item-click-event';
-export * from './list-button';
+export * from './paginable-action-button';
 export * from './paginable-table-config';
 export * from './paginable-table-dropdown';
 export * from './paginable-table-header';
@@ -9,6 +9,5 @@ export * from './paginable-table-options';
 export * from './paginable-table-ordination';
 export * from './paginable-table-pagination';
 export * from './pagination-state';
-export * from './row-button';
 export * from './table-row';
 export * from './table-row-event';